refactored to more flexible tempus.edge.track.manager namespace
