+++ applications/track-manager/pom.xml
@@ -232,7 +232,7 @@
           <finalName>${project.artifactId}-uber</finalName>
           <transformers>
             <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-              <mainClass>com.hashmapinc.tempus.edge.track_manager.Driver</mainClass>
+              <mainClass>com.hashmapinc.tempus.edge.track.manager.Driver</mainClass>
             </transformer>
           </transformers>
           <filters>
