+++ applications/track-manager/src/main/scala/com/hashmapinc/tempus/edge/track/manager/iofog/IofogController.scala
@@ -1,4 +1,4 @@
-package com.hashmapinc.tempus.edge.track_manager.iofog
+package com.hashmapinc.tempus.edge.track.manager.iofog
 
 import javax.json.JsonObject
 import play.api.libs.json.Json
@@ -6,7 +6,7 @@ import com.iotracks.api.listener.IOFogAPIListener
 import com.iotracks.elements.IOMessage
 import com.typesafe.scalalogging.Logger
 
-import com.hashmapinc.tempus.edge.track_manager.{IofogConfig, Config}
+import com.hashmapinc.tempus.edge.track.manager.{IofogConfig, Config}
 
 /**
  * This object holds the logic for handling iofog events
